backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
google-cloud-storage==2.18.2
mysql-connector-python==9.0.0